LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/cross-origin-isolated-permission.https-expected.txt


FAIL frame: origin = https://localhost:9443, value = undefined assert_equals: expected true but got false
FAIL frame: origin = https://localhost:9443, value = * assert_equals: expected true but got false
FAIL frame: origin = https://localhost:9443, value = self assert_equals: expected true but got false
PASS frame: origin = https://localhost:9443, value = (\)
PASS frame: origin = https://127.0.0.1:9443, value = undefined
PASS frame: origin = https://127.0.0.1:9443, value = *
PASS frame: origin = https://127.0.0.1:9443, value = self
PASS frame: origin = https://127.0.0.1:9443, value = (\)
FAIL dedicated worker: scheme = https, value = undefined assert_equals: expected true but got false
FAIL dedicated worker: scheme = https, value = * assert_equals: expected true but got false
FAIL dedicated worker: scheme = https, value = self assert_equals: expected true but got false
PASS dedicated worker: scheme = https, value = (\)
PASS dedicated worker: scheme = data, value = undefined
PASS dedicated worker: scheme = data, value = *
PASS dedicated worker: scheme = data, value = self
PASS dedicated worker: scheme = data, value = (\)
FAIL dedicated worker: scheme = blob, value = undefined assert_equals: expected true but got false
FAIL dedicated worker: scheme = blob, value = * assert_equals: expected true but got false
FAIL dedicated worker: scheme = blob, value = self assert_equals: expected true but got false
PASS dedicated worker: scheme = blob, value = (\)
FAIL shared worker: withCoopCoep = false Can't find variable: SharedWorker
FAIL shared worker: withCoopCoep = true Can't find variable: SharedWorker
FAIL service worker: withCoopCoep = false assert_equals: expected (boolean) false but got (undefined) undefined
FAIL service worker: withCoopCoep = true assert_equals: expected (boolean) true but got (undefined) undefined
